remove mock data in node host environment
