+++ localenv/api.js
@@ -1,10 +1,10 @@
 var express = require('express');
 var router = express.Router();
-var Salus = require('./routes/salus');
+var Resources = require('./routes/resources');
 var Metrics = require('./routes/metrics');
 
 //all Monitoring V2 routes
-router.use('/salus', Salus);
+router.use('/salus', Resources);
 
 //all Metrics V2 routes
 router.use('/metrics', Metrics);
